nxp: Correct BL2 LIMIT checks
For ls1088 and ls2088, in case of secure boot, last 16K of OCRAM
needs to be reserved for CSF header for BL2. This would mean that
load section of BL2 shouldn't exceed this boundary. Later at run
time this area would be used for SD buffer for SD boot.
The limits BL2_LIMIT and BL2_TEXT_LIMIT have been adjusted
accordingly.

Signed-off-by: Ruchika Gupta <ruchika.gupta@nxp.com>
Signed-off-by: rocket <rod.dorris@nxp.com>