nxp: Fix coherency issue for restarting core
Core status is updated when a core boot through warmboot entrypoint
with cache and MMU disabled. Other cores access the same data with
cache enabled. Even the cache is clean, it doesn't contain the up-
dated data from the booting core. The solution is to invalidate
dcache at the end of _setCoreState().

Signed-off-by: York Sun <york.sun@nxp.com>