plat/nxp: Correct calculation of GIC Offset for LS1023
There was an additional check on SVR along with
Revsion check to determine the GIC alignement offset.
The SVR checking was limited to LS1043 causing failure on
LS1023. Since this function is executed from LS1043 SoC
specific function, SVR version check is redundant. Remove
the SVR check to remove the error.

Signed-off-by: Ruchika Gupta <ruchika.gupta@nxp.com>