plat: nxp: Add OTA support with POLICY_OTA option
This patch adds OTA support with POLICY_OTA option. The OTA status
information will be read in BL2 stage. Then the Bootstrap FIP on SD card
or the Solution FIP on Flash will be loaded according to the OTA status.

OTA support is enabled with POLICY_OTA option. If you want to enable
POLICY_OTA, please add POLICY_OTA=1 in the build command.

Signed-off-by: Alison Wang <alison.wang@nxp.com>