lx2160a: timer: Configure timer0 as Non Secure
Level 1 of the SBSA requires that unless all of the local PE timers
are always on that there is a system wakeup timer in the form of the
memory mapped timer. In systems that implement EL3, the memory mapped
timer (the CNTBaseN frame and associated CNTCTLBase frame) must be
mapped into the Non-secure address space.

In lx2160a SoC, all the four timers are memory mapped.
Timer0 can be configured as a non secure timer.
To comply with SBSA Spec, mark Timer0 as Non secure timer.

Signed-off-by: Pankaj Gupta <pankaj.gupta@nxp.com>
Signed-off-by: Pramod Kumar <pramod.kumar_1@nxp.com>