for low pwr-released core frm boot-hold-off to wfe
On LX2, keeping the cores in boot-hold-off draw high
power as if all cores are in while(1);

To attain low power state, all the core are released
from boot-hold-off and put in wfe.

A scratch pad register 14 of DCFG address space is used
to store the state transition from wfe to warm-boot.

Signed-off-by: Pankaj Gupta <pankaj.gupta@nxp.com>