plat: nxp: ls1088a: fix erratum A-010194 about LPM20
Issue: There are two paths for wake request to be routed to GPP ARM
cores. The first path is from GIC to PMU. This path can be masked with
TPMWAKEMR0 register. The second path is from GIC to PMU then to EPU.
From EPU, after some control logic, it is then routed back to PMU. The
first path and second path(back from EPU) are then logically ORed and
connected to EVENTI of ARM cores. There is logic problem with the first
path for core 0, which is the last active core for LPM20.

Workaround: ARM A5x core in STANDBYWFE should not be woken up by GIC
wake request but must be woken up by EVENTI instead. This requires a
change in the software to re-target the wakeup interrupts to a core
other than the last active core0.

Signed-off-by: Ran Wang <ran.wang_1@nxp.com>