DDR driver does not check for errors, leading to undiagnosed board malfunction.
Hence, Added support to read the value of ERR_DETECT register.
If the value is non-zero the will check for the which error
has occurred during the ddr-init. For the error type other
then "Automatic calibration error" only error log will be logged.
for error type "Automatic calibration error" failure will be
returned to caller function.

Signed-off-by: Munish Kumar <munish.kumar_2@nxp.com>